完善Netty模式

+++ pcaplib/src/main/java/com/grumpycat/pcaplib/NetProxy.java
@@ -4,7 +4,7 @@ package com.grumpycat.pcaplib;
  * Created by cc.he on 2018/11/23
  */
 public interface NetProxy {
-    int getPort();
+    short getPort();
     void start();
     void stop();
 }
